# src/users/docs/swagger.yaml
basePath: /
definitions:
  presenter.HTTPError:
    properties:
      code:
        example: 400
        type: integer
      reason:
        type: object
    type: object
  user.EnrollmentRequest:
    properties:
      expire_at:
        description: Date of user's contract expiration if defined
        example: 2006-01-02T15:04:05Z07:00
        type: string
      role:
        description: Role of the user
        example: admin
        type: string
      user_id:
        description: User's unique identifier
        example: f4bc94f1-3af4-4ae0-9330-19d86ca42b30
        type: string
    required:
    - role
    - user_id
    type: object
  user.RegistrationRequest:
    properties:
      email:
        description: Email address of the user
        example: john.smith@example.com
        type: string
      firstname:
        description: First name of the user
        example: John
        type: string
      lastname:
        description: Last name of the user
        example: Smith
        type: string
    required:
    - email
    - firstname
    - lastname
    type: object
  user.User:
    properties:
      confirmed:
        description: User is confirmed by admin.
        type: boolean
      created_at:
        description: User initial registration date and time
        example: 2021-01-02T15:04:05Z07:00
        type: string
      email:
        description: Email address of the user
        example: john.smith@example.com
        type: string
      enrollment_id:
        description: User's enrollment identifier
        example: e9f437b4-2622-47c3-8d91-921e710c7354
        type: string
      expire_at:
        description: Date of user's contract expiration if defined
        example: 2021-05-02T15:04:05Z07:00
        type: string
      firstname:
        description: First name of the user
        example: John
        type: string
      id:
        description: User's unique identifier
        example: f4bc94f1-3af4-4ae0-9330-19d86ca42b30
        type: string
      lastname:
        description: Last name of the user
        example: Smith
        type: string
      role:
        description: Role of the user
        example: admin
        type: string
    type: object
host: identity.chainmetric.network
info:
  contact:
    email: timauthx@gmail.com
    name: Timothy Yalugin
    url: https://github.com/timoth-y
  description: Chainmetric Identity API users to authenticate and save preferences.
  license:
    name: Apache 2.0
    url: https://raw.githubusercontent.com/timoth-y/chainmetric-contracts/main/LICENSE
  title: Chainmetric Identity API
  version: "1.0"
paths:
  /users/enroll:
    get:
      consumes:
      - application/json
      description: Generates signing cryptographic identity for user and confirms
        one.
      parameters:
      - description: Request to enroll new user
        in: body
        name: register
        required: true
        schema:
          $ref: '#/definitions/user.EnrollmentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/presenter.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/presenter.HTTPError'
      summary: Enroll new user
      tags:
      - users
  /users/register:
    get:
      consumes:
      - application/json
      description: Request user initial registration.
      parameters:
      - description: Request to register new user
        in: body
        name: register
        required: true
        schema:
          $ref: '#/definitions/user.RegistrationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/presenter.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/presenter.HTTPError'
      summary: User registration
      tags:
      - users
swagger: "2.0"